feat(block-reverter): only require private key for sending revert transactions (#1579)
## What ❔

Previosly whenever working with `block_reverter` you needed to specify
the private key that is eneded to send the revert transaction. However
this key is only required for one operation and makes no sense while the
user just reverts a database or a rocksdb. This patch
* Makes operator private key required only for sending revert
transactions
* Makes operator _address_ required to print suggested values

Thus limiting the usage of private key to a minimum.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.