feat: Prover CLI Scaffoldings (#1609)
## What ❔

This PR implements the scaffolding for a CLI to interact with the
prover. Additionally, it integrates the functionality of the `tool`
crate.

## Why ❔

The purpose of this CLI is to simplify the process of interacting with
the prover and enhance its usability. Users can easily interact with the
prover without requiring specialized insider knowledge, ensuring that it
can managed without direct database access and complex manual queries.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.

---------

Co-authored-by: Lech <88630083+Artemka374@users.noreply.github.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>