fix: updated protobuf ci (#1803)
It turns out that the current implementation of the check compares a
given pr against the head of the base branch. If some proto files have
been updated on the base branch (so there is a fork between the pr and
the base branch), compatibility check will think that the pr wants to
revert the changes made on the base branch.

This pr fixes that by making the check compare pr against the lowest
common ancestor between the pr and the base branch.