Update package.json - Fix security vulnerability in adm-zip package
This commit updates the adm-zip package to version 0.5.2 to fix a security vulnerability related to directory traversal. The vulnerability could allow an attacker to extract files outside the target folder when processing compressed zip data. By upgrading to the latest version of adm-zip, we ensure that our codebase is protected against this exploit and maintain the security integrity of our project.