fix(flake): update build inputs and environment setup in flake.nix (#1649)
## What ❔

This change updates the cargoHash in the flake.nix file and brings in
new build dependencies: snappy.dev, lz4.dev, and bzip2.dev. Also, it
modifies the shellHook scripts to provide necessary setup for using the
`mold` linker. In preparing for these changes, the `./sdk` directory was
removed from the sources list and the `devShells` section was also
updated accordingly.

## Why ❔

This PR provides a working nix development environment, even with
existing `nix-ld` settings.

Signed-off-by: Harald Hoyer <harald@matterlabs.dev>