chore(node_framework): Fix flaky test (#1844)
## What ❔

Due to the fact that the task wrapper has a `tokio::select` [waiting for
either
](https://github.com/matter-labs/zksync-era/blob/main/core/node/node_framework/src/task.rs#L69)
barrier or stop signal, the second task in the test may not be started
at all. This is highly unlikely, but it happened at least 2 times on CI.

This PR fixes it by introducing an additional barrier that makes sure
that both tasks have started.

The fix is somewhat clunky, but it should do the job, and the tests
there warrant a refactoring anyways, so let's keep it like that for the
time being.

## Why ❔

Flaky tests bad. 

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.