feat: prover components versioning (#1660)
Use the defined protocol version on the components of prover subsystems.
Refactor queries to check the protocol version of the job they are
trying to access.

To be consistent with protocol versions.
To be able to run 2 different versions of provers at the same time with
no intersection

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.