feat(Prover CLI): Configuration (with flag) (#1861)
## What ❔

<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

Adds the possibility to configure the DB URL. For now, we use a flag
`--db-url` for passing it. If not passed it defaults to
`postgres://postgres:notsecurepassword@localhost/prover_local`.

In the future, we should handle this differently, with an independent
command `config` to set one or multiple configurations into a
`.pli-config` file.

This also removes the need for the CLI to be run with `zk f`.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

We need to offer the Prover CLI users to provide the DB URL to which
they want to work. In addition, we want to get rid of the need for `zk
f` to run the app.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.

---------

Co-authored-by: Joaquin Carletti <56092489+ColoCarletti@users.noreply.github.com>
Co-authored-by: EmilLuta <EmilLuta@users.noreply.github.com>