feat(pli): envfile loading
Load environment variables from file. Defaults to reading them from
`$ZKSYNC_HOME/etc/pliconfig`, overridable via the `PLI__CONFIG`
variable.
The parser is intentionally simple and permissive. It will simply not
load anything from lines starting with `#` (considered a comment) or not
containing an `=`.
Everything else is loaded into the environment with everything to the
left of `=` used as key and everything to the right of `=` used as value.