feat(Prover CLI): `requeue` cmd (#1719)
## What ❔

<!-- What are the changes this PR brings about? -->
<!-- Example: This PR adds a PR template to the repo. -->
<!-- (For bigger PRs adding more context is appreciated) -->

Adds a `requeue` command that requeues a given batch if it's stuck. It
looks at all jobs that are stuck and requeues them. A job is stuck if it
has a large number of attempts and is not successful.

For now, the attempts are set via a cmd flag (`--max-attempts`) which if
not set, is 10 by default.

### Usage Example

```
cd prover/prover_cli
zk f cargo run --release requeue --batch 1
```

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

We want to be able to requeue a stuck batch  with the CLI.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.

---------

Co-authored-by: Ivan Litteri <ivanlitteri@Ivans-MacBook-Pro.local>
Co-authored-by: Joaquin Carletti <joaquin.carletti@lambdaclass.com>
Co-authored-by: Joaquin Carletti <56092489+ColoCarletti@users.noreply.github.com>