refactor: Split the rest of the zksync_core (#1940)
Extracts remaining (strategically meaningful) parts of `zksync_core`.

This was a hell of a feat to de-entangle all of these interdependencies
(especially in tests, oh boy).
The following crates were extracted:

- zksync_node_test_utils
- zksync_state_keeper
- zksync_reorg_detector
- zksync_consistency_checker
- zksync_metadata_calculator
- zksync_node_sync
- zksync_node_consensus
- zksync_contract_verification_server
- zksync_node_api_server
- zksync_tee_verifier_input_producer

Contents of `zksync_core::utils` were moved to other places.
`zksync_core` was renamed to `zksync_core_leftovers` (and will be
removed when we migrate to the framework).

⚠️ this is not a perfect PR. In some places, moving stuff to another
location may not be fully idiomatic. In other places, the visibility of
many items could've been reduced. However, it was TOUGH to get this PR
even to this form, so let's focus on major issues. All the minor things
are to be polished later.

Once this is merged, I'll also update settings for infra to enable logs
for new crates.