feat(vm-runner): implement output handler for VM runner (#1856)
## What ❔

This PR creates a new abstraction `OutputHandlerFactory` that can
orchestrate output handlers in a non-trivial manner. Specifically, for
the purposes of VM runner we create an implementation
`ConcurrentOutputHandlerFactory` that can produce delegated output
handlers that internally use another output handler (i.e. delegatee is
specific to the VM runner implementation). The main benefit of this
implementation is that `handle_l1_batch` is non-blocking and finishes
almost immediately. The actual work is spawned as a future and is then
handled asynchronously enabling us to create many output handlers in
parallel.

## Why ❔

One of the components for the upcoming VM runner

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.