Make protocol upgrade tool support semver (#2081)
## What ❔

We'll store the semantic version of the protocol in the `common.json`.
For the upgrade itself we'll have to use packed format, but for the l2
transaction we'll use only the `minor` version (for compatibility with
the server). It is expected that the upgrade transaction is only present
during at least `minor` upgrades and never present during `patch` ones

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.