feat(node_framework): Synchronize pools layer with logic in initialize_components (#2079)
## What ❔

There was some new logic related to the configuration of DAL/pools in
`initialize_components` that wasn't mirrored in the pools layer.
This PR changes the pools layer to match logic in
`initialize_components`.

## Why ❔

We do not intend the framework to change the way the server works.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.