fix(api): correct default fee data in eth call (#2072)
## What ❔

Previously, if no `max_fee_per_gas` was provided, it would mean that we
would use 0 as the base fee. This would also require the gas per pubdata
to be 0.

With this PR we will now use the current gas per pubdata. This PR also
provides a large eth call gas limit to ensure that it will work fine
even under very high L1 gas price

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.