fix(vm-runner): add config value for the first processed batch (#2158)
## What ❔

Adds a new config value that regulate which batch to consider as the
last processed on a fresh start.

Also renames config values to be more idiomatic due to several from
DevOps and other people :)

## Why ❔

We don't want to wait half a year for VM to process all testnet/mainnet
batches.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.