fix(prover_cli): Fix the issues with `home` path (#2104)
In this PR: [Pull Request
#2022](https://github.com/matter-labs/zksync-era/pull/2022/files), the
logic was changed to stop using the $ZKSYNC_HOME environment variable to
construct paths relative to the root of zksync-era.
But the prover is a separate workspace, so it fails to create the path
to the contract with the functions in the main workspace..

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.