fix(api): Fix getting pending block (#2186)
## What ❔

Fixes getting data for the pending block in `eth_getBlockByNumber` and a
couple of other methods.

## Why ❔

Getting a pending block should always return `null`, but right now it
actually doesn't. Caused by non-atomic reads from Postgres in the method
implementation: first, a pending block number is resolved, and then a
block with this number is fetched. Between these two reads, a block with
this number may be inserted to Postgres.

While it's somewhat unlikely that anyone will query a pending block,
it's still a correctness issue.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.