feat(merkle-tree): Rework tree rollback (#2207)
## What ❔

Reworks tree rollback so that it's supported on the distributed external
node, including the case when a node runs multiple trees.

- Desync with Postgres is now detected on metadata calculator
initialization, and the tree is truncated correspondingly.
- The old approach is left intact as a safety guard.

## Why ❔

Right now, reorg logic on EN relies on a node running a single tree, and
block reverter being a singleton. Both these assumptions are bogus in
case of a distributed EN.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.