fix(dal): Fix contract_address field in getTransactionRecipt (#2223)
## What ❔

+ Change `get_transaction_receipts` DAL method to fetch the contract
address from the `events` table.
+ Add integration test that checks we return the correct contract
address in the receipt of the transaction that deployed inner-outer
contract(e.g. contract that initializes another contract in its
constructor).

## Why ❔
The current approach(going to the `storage_logs` table to get the
contract address) is error-prone. The problem here is that we
deduplicate storage logs, which renders operation_number invalid. So,
the ordering is not correct, and we might return the incorrect deployed
address. This is important when one tx deploys a contract that
initializes another contract in the constructor(which means we deploy
several contracts here) - the user expects the "outer" contract address
but gets the inner.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.