feat: verification of L1Batch witness (BFT-471) (#2019)
Verification of an L1 batch, based on StoredBatchInfo. We extract the
hash of the last block of the transaction via merkle path from the
root_state (for batch n and n-1) and we recompute the rolling block hash
based on transactions in the payload. The verification is not perfect as
there are still some fields in payload for which we don't have a
commitment - we should address that later.

Verification will be used for syncing L1Batches over p2p network (not
implemented yet).

Also removed serde serialization from TransactionRequest which is unused
- we don't need to maintain encoding compatibility for it.