perf(db): Improve storage switching for state keeper cache (#2234)
## What ❔

Improves switching logic between Postgres and RocksDB for SK cache. With
these changes, RocksDB is guaranteed to be used if it's up to date when
it's opened.

## Why ❔

Previously, Postgres could be used after node start (primarily if
there's a pending L1 batch) even if RocksDB is up to date. This is
caused by potential delays when opening RocksDB. This behavior was
observed in the wild, e.g. for a mainnet full node with pruning enabled.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.