fix(merkle-tree): Change `LazyAsyncTreeReader::wait()` signature (#2314)
## What ❔

Makes `LazyAsyncTreeReader::wait()` return an `Option` to make it
clearer that not initializing is not necessarily an error (this decision
is up to the caller).

## Why ❔

Recent errors on ENs were caused by unwrapping
`LazyAsyncTreeReader::wait()`. They partially masked the real error
cause (not related to this task).

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.