feat(eth-sender): handle transactions for different operators separately to increase throughtput (#2341)
Before this PR, all transactions to be send for the first time landed in
one queue.
Whenever a transaction had to be resent, on each block: 
- we chose the first non-blob transaction could a blob transaction be
resent.
- If and only if there are no such non-blob txs we resend a blob
transaction.

While this works, it's not ideal as those two types of transactions can
be safely sent independently to increase throughput.
It's also harder to think about time in mempool now as for instance a
blob txs may have to wait for an hour before all pending non-blob txs
are sent and confirmed.
My PR changes so that those two types of transaction are completely
separate. It makes sense as they are using different sender accounts
(operators) and a different set of nonces, so they can be safely sent in
parallel.

---------

Signed-off-by: tomg10 <lemures64@gmail.com>