feat: Validium with DA (#2010)
## What ❔

This PR adds an opportunity to use external DA layers to submit the
pubdata.
The implementations of the clients will be created in this repo:
https://github.com/matter-labs/hyperchain-da
This is only the first stage, and doesn't include any modifications to
the `commitBatches` or bootloader memory, those will be added as a
follow-up.
Design doc for the feature (might be a bit outdated, but main concepts
are valid):
https://matterlabs.notion.site/Validium-with-DA-EXTERNAL-8deccba433be4ff88592a3b0f8774062
It is assumed that the pubdata publishing is subsidized by the operators
at this points, it shouldn't be a concern because:
- the DA layers' storage costs are quite low
- for most DA layers, the blob size is around 2MB (or they work on
increasing it to be like that), so we could in theory fit ~30-35k txs
there, which makes that cost even lower per tx

## Why ❔

We have many partners who want to use the different DA layers, this PR
provides a common interface for all of them

## Checklist

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.