feat: Minimal External API Fetcher (#2383)
## What ❔

For chains running with a custom base token, use live prices in creating
ETH<->BaseToken conversion ratios.
This PR
- adds a first `PriceAPIClientResource` with a `PriceAPIClient`
CoinGecko client.
- Uses this client in the `BaseTokenRatioPersister` to fetch new prices
and create new ratios

Not included in this PR and will be follow-up work:
- Redundancy in the number of external price feeds used
- Different "strategies" for how true price converged from multiple
price feeds

## Why ❔

For the base token flow, we need to be be able to fetch live prices of
BaseToken & ETH from which to create the conversion ratios.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Co-authored-by: dimazhornyk <dima.zhornyk@gmail.com>
Co-authored-by: Igor Aleksanov <popzxc@yandex.ru>