feat: Update to consensus 0.1.0-rc.4 (BFT-486) (#2475)
## What ❔

Updates the `era-consensus` dependency to 0.1.0-rc.4, which brings the
following changes:
* `BatchStoreState::last` no longer contains the entire `SyncBatch`,
just the number of the latest batch, which should reduce the unintended
gossip data volume
* Makes new batch gossip related metrics available to Prometheus
(BFT-486)

## Why ❔

Adding attesters to the genesis is a breaking change, so we would like
to cut a new release to external node operators. These changes make the
gossip less resource intensive as well as make it more observable.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Co-authored-by: Bruno França <bruno@franca.xyz>