fix(proof_data_handler): Unlock jobs on transient errors (#2486)
## What ❔

Currently, proof data handler locks the job for proof generation, and
starts fetching required data after that.
If any error happens during fetching, the method will err, and the job
will remain locked.

This PR changes it, so that if any error occurs, we unlock the job
before we return an error.

Additionally, it reduces the amount of non-necessary panics in the
touched code, and adds some docs.

## Why ❔

Correctness & efficiency.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.