feat: Allow tracking l2 fees for L2-based chains (#2563)
## What ❔

In this PR we add changes to the GasAdjuster needed to track the fees on
L2-based chains. The PR aims to preserve type-safety for places that
should work when L1 interface is available only, so `DynClient<L1>` is
left in a lot of places and generally `L1` is the default network to be
used. However, some methods were extended to work with any network.

Overall what was added:

- A new resource for accessing L2-specific endpoints.
- A new field is returned from `fee_history`. Clients should not rely on
this field. However, we added it so we could test things.
- Some refactoring
- While some new new config fields are added into the Rust code, they
are not added to the config so that in case something has to be changed,
we do not introduce legacy (esp. in protobuf)
- Into protobuf we only added `RelayedL2Calldata` pubdata type, which
would symbolize that the calldata is being relayed to L1 through another
L2. I am pretty confident that this type will be needed anyway.

## Why ❔ 

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.