feat(prover): Add ProverJobMonitor (#2666)
ProverJobMonitor will be house keeper's counter part in prover
subsystem. TL;DR; it's a singleton component, monitoring prover
subsystem jobs.

The TL;DR; is that prover and core won't share any databases. This
enables:
- core deployments without affecting prover
- removing prover infrastructure (DB) in proverless envs

The release plan is as follows:
- release a component (PJM) that runs in parallel with HK
- migrate all jobs/metrics/dashboards to PJM
- delete their counterparts in HK
- remove redundant infrastructure

This PR contains:
- a new component (PJM)
- fixes for bugs/issues with old metrics (backported to HK)
- refactoring of metrics (PJM metrics cover same metrics as HK, but they
are different, as we can cover more with less)
- various other small nits

P.S. Name is up for discussion, feel free to suggest better name.