feat: update base token rate on L1 (#2589)
Currently base token to ETH ratio is always locked to 1:1 on the L1.
This PR changes that and adds functionality to update L1 numerator and
denominator as soon as they get updated on the L2. Updates are done by
calling `IChainAdmin.setTokenMultiplier` function. The function can only
be called by a new role - `tokenMultiplierSetter`.

This PR introduces the following changes:
* new wallet for `tokenMultiplierSetter`, provisioned with ETH on
network initialisation;
* updates to old and new tooling to fully support this new role;
* updates to `base_token_ratio_persister` to do the L1 updates;

What is not included:
* Conditional updates - to save gas cost if numerator / denominator
hasn't changed since the last time;
* Support for fluctuating numerator / denominator in EN - that will be
added in a separate PR.

[PR](https://github.com/matter-labs/era-contracts/pull/683/files) to
update tolling on the `era-contracts` side.