chore: Reverts prover job metrics to HK metrics
The point of this PR is to keep metrics history the same  as house
keeper such that analytics can be ran retroactively on top of it.

Whilst I find it both confusing (house keeper metrics in prover job monitor)
and of overall worst quality, I disagreed and commited with the decision.