fix(vm): Fix used bytecodes divergence (#2741)
## What ❔

Fixes divergence in used bytecodes info between the old and new VMs.

## Why ❔

The new VM behaved differently to the old VM on far call if decommitting
the called contract leads to out-of-gas revert. The old VM records the
called contract bytecode as decommitted in this case; the new one
didn't.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.