feat: make `to` address optional for transaction data (#2852)
## What ❔

This PR is the first part of [this
one](https://github.com/matter-labs/zksync-era/pull/2631).

We decided to separate it into smaller PRs to make it easier to read and
review.

This PR contains only the changes to the `TransactionData` structs for
the different VM versions, modifying them for an optional field.

The other parts are currently available in our repo fork, as they are
based on this branch. The other PRs are:
- [feat: evm simulator as base system
contract](https://github.com/lambdaclass/zksync-era/pull/256):
integrates the EVM interpreter contract as a base system contract and
includes all necessary updates for zksync-era to run with it.
- [feat: make evm simulator optional for the
server](https://github.com/lambdaclass/zksync-era/pull/257): makes the
evm simulator optional to run using a config file.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.