chore(prover): Fixes broken non-GPU setup (#2953)
Previously, this was done via a compile flag. This was deemed poor taste
and we've moved a specific env-var - `ZKSYNC_USE_CUDA_STUBS`. This is
wired deep within crypto dependencies.

This PR updates the dependencies, adds documentation on how to work with
the flag and updates all workflows (fixing the broken ones from
RUSTFLAGS time).

---------

Co-authored-by: Igor Aleksanov <popzxc@yandex.ru>