fix(tee_verifier): correctly initialize storage for re-execution
Previously the storage for VM re-execution was initialized just from `WitnessInputMerklePaths`.
This although misses the storage values for slots, which are only read/written to by rolled back transactions.

With this commit, the TEE verifier uses `WitnessStorageState` of `VMRunWitnessInputData` to initialize the storage.
This requires waiting for the BasicWitnessInputProducer to complete
and therefore the TEE verifier input producer can be removed.
The input for the TEE verifier is now collected in the `proof_data_handler`, which enables to remove
the whole job queue for the TEE verifier input producer.

Co-authored-by: Patrick Beza <pab@matterlabs.dev>
Signed-off-by: Harald Hoyer <harald@matterlabs.dev>