fix(zk_toolbox): wrong configs path for prover binaries (#2922)
## What ❔

Right now, while running prover binaries inside of zksync-era, config
paths are specified as path to config from chain config(which is
`./chain/configs` - relative path, but not absolute). This breaks,
because to run prover binaries, shell enters prover directory of era.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.