feat(zk_toolbox): improved wallet handling (#3058)
Made the tool fail with an error when wallets.yaml has parsing errors,
rather than displaying "not found".
Made the tool fail when the wallets.yaml is unparseable instead of
blindly overwriting it (in Localhost mode).
Made the tool check that the address actually matches the private key
when parsing the config.
Made the tool check that the private key is actually a valid private key
when parsing the config.
Made the code pass around Wallet, rather than weakly typed H256.