fix(tee_verifier): correctly initialize storage for re-execution (#3017)
## What ❔

With this commit, the TEE verifier uses `WitnessStorageState` of
`VMRunWitnessInputData` to initialize the storage. This requires waiting
for the BasicWitnessInputProducer to complete and therefore the TEE
verifier input producer can be removed. The input for the TEE verifier
is now collected in the `proof_data_handler`, which enables to remove
the whole job queue for the TEE verifier input producer.

## Why ❔

Previously the storage for VM re-execution was initialized just from
`WitnessInputMerklePaths`. This although misses the storage values for
slots, which are only read/written to by rolled back transactions. This
led to failed verification of blocks, which would normally pass.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk_supervisor fmt` and `zk_supervisor
lint`.

---------

Signed-off-by: Harald Hoyer <harald@matterlabs.dev>
Co-authored-by: Patrick Beza <pab@matterlabs.dev>